relocate types from CodeGen to CodeGen.Type
