+++ hibiscus.cabal
@@ -1,6 +1,6 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.37.0.
+-- This file has been generated from package.yaml by hpack version 0.35.2.
 --
 -- see: https://github.com/sol/hpack
 
@@ -19,6 +19,7 @@ library
       Hibiscus.Lexer
       Hibiscus.Parser
       Hibiscus.CodeGen.Type.DataType
+      Hibiscus.CodeGen.Type
       Hibiscus.CodeGen
       Hibiscus.Asm
       Hibiscus.Type4plus
